+++ src/db/Migration.php
@@ -238,12 +238,6 @@ abstract class Migration extends \yii\db\Migration
      */
     public function upsert($table, $insertColumns, $updateColumns = true, $params = [], bool $includeAuditColumns = true): void
     {
-        if (is_bool($params)) {
-            $includeAuditColumns = $params;
-            $params = [];
-            Craft::$app->getDeprecator()->log('craft\\db\\Migration::upsert($includeAuditColumns)', 'The `$includeAuditColumns` argument on `craft\\db\\Migration::upsert()` has been moved to the 5th position');
-        }
-
         $time = $this->beginCommand("upsert into $table");
         $this->db->createCommand()->upsert($table, $insertColumns, $updateColumns, $params, $includeAuditColumns)->execute();
         $this->endCommand($time);